file moving for constants, utils, pages
