+++ src/actions/PageActions.js
@@ -1,4 +1,4 @@
-import { SET_EVENTS, SET_EVENTS_REGISTERED } from '../constants/Constants'
+import { SET_EVENTS, SET_EVENTS_REGISTERED } from 'constants/index'
 
 export function setEvents (events) {
   return {
