+++ src/components/Nav.scss
@@ -1,4 +1,4 @@
-@import '../constants/Constants.scss';
+@import '../constants/index.scss';
 
 .Navigation {
     height: 100%;
