+++ src/components/Router.scss
@@ -1,4 +1,4 @@
-@import '../constants/Constants.scss';
+@import '../constants/index.scss';
 
 .content {
     padding-top: 14px;
