+++ src/components/icons/BiztechIcon.js
