+++ src/components/icons/LogoutIcon.js
