--- src/constants/Constants.js
@@ -1,18 +1,0 @@
-export const SET_EVENTS = 'SET_EVENTS'
-export const SET_EVENTS_REGISTERED = 'SET_EVENTS_REGISTERED'
-export const SET_USER = 'SET_USER'
-export const LOGOUT = 'LOGOUT'
-export const REGISTRATION_STATUS = {
-  REGISTERED: 'registered',
-  CHECKED_IN: 'checkedIn',
-  WAITLISTED: 'waitlist',
-  CANCELLED: 'cancelled'
-}
-export const COLOR = {
-  WHITE: '#FFFFFF',
-  FONT_COLOR: '#AEC4F4',
-  CARD_PAPER_COLOR: '#273149',
-  BACKGROUND_COLOR: '#070F21',
-  LIGHT_BACKGROUND_COLOR: '#172037',
-  BIZTECH_GREEN: '#7AD040'
-}
