+++ src/constants/_constants/actions.js
@@ -1,0 +1,4 @@
+export const SET_EVENTS = 'SET_EVENTS'
+export const SET_EVENTS_REGISTERED = 'SET_EVENTS_REGISTERED'
+export const SET_USER = 'SET_USER'
+export const LOGOUT = 'LOGOUT'
