+++ src/constants/_constants/registration.js
@@ -1,0 +1,6 @@
+export const REGISTRATION_STATUS = {
+  REGISTERED: 'registered',
+  CHECKED_IN: 'checkedIn',
+  WAITLISTED: 'waitlist',
+  CANCELLED: 'cancelled'
+}
