+++ src/constants/_constants/theme.js
@@ -1,0 +1,8 @@
+export const COLORS = {
+  WHITE: '#FFFFFF',
+  FONT_COLOR: '#AEC4F4',
+  CARD_PAPER_COLOR: '#273149',
+  BACKGROUND_COLOR: '#070F21',
+  LIGHT_BACKGROUND_COLOR: '#172037',
+  BIZTECH_GREEN: '#7AD040'
+}
