+++ src/constants/index.js
@@ -1,0 +1,7 @@
+export * from './_constants/actions'
+
+export * from './_constants/config'
+
+export * from './_constants/registration'
+
+export * from './_constants/theme'
