+++ src/constants/index.scss
@@ -1,0 +1,1 @@
+@import "./_constants/theme.scss"
