--- src/onboarding.txt
@@ -1,6 +1,0 @@
-add your name here guys:
-derek
-merissa
-patrick 
-marco
-sami
