+++ src/pages/admin/Event/EventCreate.js
@@ -1,15 +1,16 @@
 import React, { useState } from 'react'
-import { useParams, useHistory } from 'react-router-dom'
+import { Helmet } from 'react-helmet'
 import * as Yup from 'yup'
 import { Formik } from 'formik'
-import EditEventForm from '../../components/Forms/EditEvent'
-import EventView from '../../components/EventView'
-import { fetchBackend, updateEvents } from '../../utils'
-import { Helmet } from 'react-helmet'
-import { makeStyles } from '@material-ui/core/styles'
 
+import EventView from 'components/EventView'
+import NewEventForm from 'components/Forms/NewEvent'
+
+import { makeStyles } from '@material-ui/core/styles'
 import Paper from '@material-ui/core/Paper'
 import Typography from '@material-ui/core/Typography'
+
+import { fetchBackend, log } from 'utils'
 
 const useStyles = makeStyles(theme => ({
   layout: {
@@ -29,31 +30,18 @@ const useStyles = makeStyles(theme => ({
   }
 }))
 
-function EventEdit (props) {
+export default function EventCreate () {
   const classes = useStyles()
-  const { id: eventId } = useParams()
-  const [event, setEvent] = useState(null)
   const [previewEvent, setPreviewEvent] = useState({})
-  const history = useHistory()
-
-  const { events } = props
-  if (!events) {
-    updateEvents()
-  }
-
-  // Get the initial values
-  if (!event && events && eventId) {
-    const event = events.find(event => event.id === eventId)
-    setEvent(event)
-    setPreviewEvent(event)
-  }
 
   const validationSchema = Yup.object({
     ename: Yup.string().required(),
+    slug: Yup.string().matches(/^[a-z\-0-9]*$/, 'Slug must be lowercase and have no whitespace').required(),
     description: Yup.string().required(),
     capacity: Yup.number('Valid number required')
       .min(0, 'Valid capacity required')
       .required(),
+    // partners: Yup.string().required(),
     elocation: Yup.string().required(),
     longitude: Yup.number('Valid number required')
       .min(-180, 'Valid number required')
@@ -67,20 +55,9 @@ function EventEdit (props) {
     imageUrl: Yup.string().url().required()
   })
 
-  const initialValues = event ? {
-    ename: event.ename,
-    slug: event.id,
-    description: event.description,
-    capacity: event.capac,
-    facebookUrl: event.facebookUrl,
-    elocation: event.elocation || '',
-    longitude: event.longitude || '',
-    latitude: event.latitude || '',
-    imageUrl: event.imageUrl,
-    startDate: event.startDate,
-    endDate: event.endDate
-  } : {
+  const initialValues = {
     ename: '',
+    slug: '',
     description: '',
     capacity: '',
     facebookUrl: '',
@@ -88,26 +65,26 @@ function EventEdit (props) {
     longitude: '',
     latitude: '',
     imageUrl: '',
-    startDate: '',
-    endDate: ''
+    startDate: new Date(),
+    endDate: new Date()
   }
 
-  return event && (
+  return (
     <div className={classes.layout}>
       <Helmet>
-        <title>Edit {event.ename} - BizTech Admin</title>
+        <title>Create Event - BizTech Admin</title>
       </Helmet>
       <Paper className={classes.paper}>
         <div className={classes.content}>
           <Typography variant='h4' align='center' gutterBottom>
-                        Edit Event
+                        New Event
           </Typography>
           <Formik
             initialValues={initialValues}
             validationSchema={validationSchema}
             onSubmit={submitValues}
           >
-            {props => <EditEventForm updatePreview={setPreviewEvent} {...props} />}
+            {props => <NewEventForm updatePreview={setPreviewEvent} {...props} />}
           </Formik>
         </div>
       </Paper>
@@ -121,6 +98,7 @@ function EventEdit (props) {
   async function submitValues (values) {
     const body = {
       ename: values.ename,
+      id: values.slug,
       description: values.description,
       capac: values.capacity,
       elocation: values.elocation,
@@ -131,17 +109,16 @@ function EventEdit (props) {
       startDate: values.startDate,
       endDate: values.endDate
     }
-
-    fetchBackend(`/events/${values.slug}`, 'PATCH', body)
-      .then((response) => {
-        alert(response)
-        history.push(`/event/${values.slug}/register`)
+    fetchBackend('/events', 'POST', body)
+      .then(response => {
+        alert(response.message)
+        window.location.href = '/'
       })
       .catch(err => {
-        console.log(err)
-        alert(err.message + ' Please contact a dev')
+        log(err)
+        if (err.status === 409) {
+          alert('Failed. Event with that slug/id already exists')
+        } else alert(err.message + ' Please contact a dev')
       })
   }
 }
-
-export default EventEdit
