+++ src/pages/public/Event/EventRegister.js
@@ -1,17 +1,20 @@
 import React, { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { connect } from 'react-redux'
-import { makeStyles } from '@material-ui/core/styles'
-import { fetchBackend, updateEvents } from '../../utils'
+import { Helmet } from 'react-helmet'
 import * as Yup from 'yup'
 import { Formik } from 'formik'
-import RegisterEvent from '../../components/Forms/RegisterEvent'
+
+import EventView from 'components/EventView'
+import RegisterEvent from 'components/Forms/RegisterEvent'
+import NotFound from 'pages/NotFound'
+
 import Paper from '@material-ui/core/Paper'
 import Grid from '@material-ui/core/Grid'
 import Skeleton from '@material-ui/lab/Skeleton'
-import { Helmet } from 'react-helmet'
-import EventView from '../../components/EventView'
-import NotFound from '../NotFound'
+import { makeStyles } from '@material-ui/core/styles'
+
+import { fetchBackend, updateEvents } from 'utils'
 
 const useStyles = makeStyles(theme => ({
   layout: {
