+++ src/utils/_utils/log.js
@@ -1,0 +1,5 @@
+export function log (message) {
+  if (process.env.REACT_APP_STAGE !== 'production') {
+    console.log(message)
+  }
+}
